Finalisation plugins géométrie

Closes #623

See merge request 3d/cars!577

+++ cars/core/geometry/otb_geometry.py
@@ -38,6 +38,7 @@ from cars.core import inputs
 from cars.core.geometry import AbstractGeometry
 from cars.core.otb_adapters import encode_to_otb
 from cars.core.utils import get_elevation_range_from_metadata
+from cars.data_structures import cars_dataset
 
 
 @AbstractGeometry.register_subclass("OTBGeometry")
@@ -52,7 +53,7 @@ class OTBGeometry(AbstractGeometry):
     @property
     def conf_schema(self):
         """
-        Returns the input configuration fields required by the geometry loader
+        Returns the input configuration fields required by the geometry plugin
         as a json checker schema. The available fields are defined in the
         cars/conf/input_parameters.py file
 
@@ -79,51 +80,31 @@ class OTBGeometry(AbstractGeometry):
         return not_available
 
     @staticmethod
-    def check_products_consistency(cars_conf) -> bool:
-        """
-        Test if the product is readable by the OTB
-
-        :param: cars_conf: cars input configuration dictionary
-        :return: True if the products are readable, False otherwise
-        """
-        # get images paths
-        img1 = cars_conf[input_parameters.IMG1_TAG]
-        img2 = cars_conf[input_parameters.IMG2_TAG]
-
-        # test if both images have associated RPC
-        status = False
-        if OTBGeometry.check_geom_consistency(
-            img1
-        ) and OTBGeometry.check_geom_consistency(img2):
-            status = True
-
-        return status
-
-    @staticmethod
-    def check_geom_consistency(img: str) -> bool:
+    def check_product_consistency(sensor: str, geomodel: str = None) -> bool:
         """
         Check if the image have RPC information readable by the OTB
-
-        :param img: path to the image
+        :param sensor: path to the image
+        :param geomodel: path and attributes for geometric model
         :return: True if the RPC are readable, False otherwise
         """
         can_open_status = False
         try:
-            geom_path = "./otb_can_open_test.geom"
+            if geomodel is None:
+                geom_path = "./otb_can_open_test.geom"
 
-            # try to dump .geom with ReadImageInfo app
-            read_im_app = otbApplication.Registry.CreateApplication(
-                "ReadImageInfo"
-            )
-            read_im_app.SetParameterString("in", img)
-            read_im_app.SetParameterString("outkwl", geom_path)
+                # try to dump .geom with ReadImageInfo app
+                read_im_app = otbApplication.Registry.CreateApplication(
+                    "ReadImageInfo"
+                )
+                read_im_app.SetParameterString("in", sensor)
+                read_im_app.SetParameterString("outkwl", geom_path)
 
-            read_im_app.ExecuteAndWriteOutput()
-
+                read_im_app.ExecuteAndWriteOutput()
+            else:
+                geom_path = geomodel["path"]
             # check geom consistency
             if os.path.exists(geom_path):
                 can_open_status = True
-
                 with open(geom_path, encoding="utf-8") as geom_file_desc:
                     geom_dict = {}
                     for line in geom_file_desc:
@@ -153,17 +134,17 @@ class OTBGeometry(AbstractGeometry):
                             )
                         )
                         can_open_status = False
-
-                os.remove("./otb_can_open_test.geom")
+                if geomodel is None:
+                    os.remove("./otb_can_open_test.geom")
             else:
                 logging.warning(
-                    "{} does not have associated geom file".format(img)
+                    "{} does not have associated geom file".format(sensor)
                 )
                 can_open_status = False
         except Exception as read_error:
             logging.warning(
                 "Exception caught while trying to read file {}: {}".format(
-                    img, read_error
+                    sensor, read_error
                 )
             )
             can_open_status = False
@@ -171,7 +152,10 @@ class OTBGeometry(AbstractGeometry):
 
     @staticmethod
     def triangulate(
-        cars_conf,
+        sensor1,
+        sensor2,
+        geomodel1,
+        geomodel2,
         mode: str,
         matches: Union[xr.Dataset, np.ndarray],
         grid1: str,
@@ -181,23 +165,29 @@ class OTBGeometry(AbstractGeometry):
         """
         Performs triangulation from cars disparity or matches dataset
 
-        :param cars_conf: cars input configuration dictionary
+        :param sensor1: path to left sensor image
+        :param sensor2: path to right sensor image
+        :param geomodel1: path and attributes for left geomodel
+        :param geomodel2: path and attributes for right geomodel
         :param mode: triangulation mode
                (cst.DISP_MODE or cst.MATCHES)
         :param matches: cars disparity dataset or matches as numpy array
-        :param grid1: path to epipolar grid of image 1
-        :param grid2: path to epipolar grid of image 2
+        :param grid1: path or dataset for epipolar grid of image 1
+        :param grid2: path or dataset for epipolar grid of image 2
         :param roi_key: dataset roi to use
                (can be cst.ROI or cst.ROI_WITH_MARGINS)
         :return: the long/lat/height numpy array in output of the triangulation
         """
 
-        img1 = cars_conf[input_parameters.IMG1_TAG]
-        img2 = cars_conf[input_parameters.IMG2_TAG]
-
         # Retrieve elevation range from imgs
-        (min_elev1, max_elev1) = get_elevation_range_from_metadata(img1)
-        (min_elev2, max_elev2) = get_elevation_range_from_metadata(img2)
+        (min_elev1, max_elev1) = get_elevation_range_from_metadata(sensor1)
+        (min_elev2, max_elev2) = get_elevation_range_from_metadata(sensor2)
+
+        # get path if grid is of type CarsDataset
+        if isinstance(grid1, cars_dataset.CarsDataset):
+            grid1 = grid1.attributes["path"]
+        if isinstance(grid2, cars_dataset.CarsDataset):
+            grid2 = grid2.attributes["path"]
 
         # Build triangulation app
         triangulation_app = otbApplication.Registry.CreateApplication(
@@ -243,8 +233,8 @@ class OTBGeometry(AbstractGeometry):
 
         triangulation_app.SetParameterString("leftgrid", grid1)
         triangulation_app.SetParameterString("rightgrid", grid2)
-        triangulation_app.SetParameterString("leftimage", img1)
-        triangulation_app.SetParameterString("rightimage", img2)
+        triangulation_app.SetParameterString("leftimage", sensor1)
+        triangulation_app.SetParameterString("rightimage", sensor2)
         triangulation_app.SetParameterFloat("leftminelev", min_elev1)
         triangulation_app.SetParameterFloat("leftmaxelev", max_elev1)
         triangulation_app.SetParameterFloat("rightminelev", min_elev2)
@@ -256,12 +246,12 @@ class OTBGeometry(AbstractGeometry):
 
         return llh
 
-    @staticmethod
     def generate_epipolar_grids(
-        cars_conf,
-        dem: Union[None, str] = None,
-        geoid: Union[None, str] = None,
-        default_alt: Union[None, float] = None,
+        self,
+        sensor1,
+        sensor2,
+        geomodel1,
+        geomodel2,
         epipolar_step: int = 30,
     ) -> Tuple[
         np.ndarray, np.ndarray, List[float], List[float], List[int], float
@@ -269,10 +259,10 @@ class OTBGeometry(AbstractGeometry):
         """
         Computes the left and right epipolar grids
 
-        :param cars_conf: cars input configuration dictionary
-        :param dem: path to the dem folder
-        :param geoid: path to the geoid file
-        :param default_alt: default altitude to use in the missing dem regions
+        :param sensor1: path to left sensor image
+        :param sensor2: path to right sensor image
+        :param geomodel1: path and attributes for left geomodel
+        :param geomodel2: path and attributes for right geomodel
         :param epipolar_step: step to use to construct the epipolar grids
         :return: Tuple composed of :
 
@@ -295,22 +285,21 @@ class OTBGeometry(AbstractGeometry):
             del os.environ["OTB_GEOID_FILE"]
 
         # create OTB application
-        img1 = cars_conf[input_parameters.IMG1_TAG]
-        img2 = cars_conf[input_parameters.IMG2_TAG]
-
         stereo_app = otbApplication.Registry.CreateApplication(
             "StereoRectificationGridGenerator"
         )
 
-        stereo_app.SetParameterString("io.inleft", img1)
-        stereo_app.SetParameterString("io.inright", img2)
+        stereo_app.SetParameterString("io.inleft", sensor1)
+        stereo_app.SetParameterString("io.inright", sensor2)
         stereo_app.SetParameterInt("epi.step", epipolar_step)
-        if dem is not None:
-            stereo_app.SetParameterString("epi.elevation.dem", dem)
-        if default_alt is not None:
-            stereo_app.SetParameterFloat("epi.elevation.default", default_alt)
-        if geoid is not None:
-            stereo_app.SetParameterString("epi.elevation.geoid", geoid)
+        if self.dem is not None:
+            stereo_app.SetParameterString("epi.elevation.dem", self.dem)
+        if self.default_alt is not None:
+            stereo_app.SetParameterFloat(
+                "epi.elevation.default", self.default_alt
+            )
+        if self.geoid is not None:
+            stereo_app.SetParameterString("epi.elevation.geoid", self.geoid)
         stereo_app.Execute()
         # OTB doesn't do a new line, force it for next logger seen by user
         print("\n")
@@ -328,14 +317,14 @@ class OTBGeometry(AbstractGeometry):
             stereo_app.GetParameterFloat("epi.baseline"),
         )
 
-        origin = stereo_app.GetImageOrigin("io.outleft")
-        spacing = stereo_app.GetImageSpacing("io.outleft")
+        origin = np.copy(stereo_app.GetImageOrigin("io.outleft"))
+        spacing = np.copy(stereo_app.GetImageSpacing("io.outleft"))
 
         # Convert epipolar size depending on the pixel size
         # TODO: remove this patch when OTB issue
         # https://gitlab.orfeo-toolbox.org/orfeotoolbox/otb/-/issues/2176
         # is resolved
-        with rio.open(img1, "r") as rio_dst:
+        with rio.open(sensor1, "r") as rio_dst:
             pixel_size_x, pixel_size_y = (
                 rio_dst.transform[0],
                 rio_dst.transform[4],
@@ -361,16 +350,13 @@ class OTBGeometry(AbstractGeometry):
             disp_to_alt_ratio,
         )
 
-    @staticmethod
     def direct_loc(
-        cars_conf,
-        product_key: str,
+        self,
+        sensor,
+        geomodel,
         x_coord: float,
         y_coord: float,
         z_coord: float = None,
-        dem: str = None,
-        geoid: str = None,
-        default_elevation: float = None,
     ) -> np.ndarray:
         """
         For a given image point, compute the latitude, longitude, altitude
@@ -381,17 +367,11 @@ class OTBGeometry(AbstractGeometry):
 
         Advice: to be sure, use x,y,z inputs only
 
-        :param cars_conf: cars input configuration dictionary
-        :param product_key: input_parameters.PRODUCT1_KEY or
-               input_parameters.PRODUCT2_KEY to identify which geometric model
-               shall be taken to perform the method
+        :param sensor: path to sensor image
+        :param geomodel: path and attributes for geometrical model
         :param x_coord: X Coordinate in input image sensor
         :param y_coord: Y Coordinate in input image sensor
         :param z_coord: Z Altitude coordinate to take the image
-        :param dem: if z not defined, take this DEM directory input
-        :param geoid: if z and dem not defined, take GEOID directory input
-        :param default_elevation: if z, dem, geoid not defined, take default
-               elevation
         :return: Latitude, Longitude, Altitude coordinates as a numpy array
         """
         # save os env
@@ -404,27 +384,22 @@ class OTBGeometry(AbstractGeometry):
             )
             del os.environ["OTB_GEOID_FILE"]
 
-        # create OTB application
-        img = cars_conf[
-            input_parameters.create_img_tag_from_product_key(product_key)
-        ]
-
         s2c_app = otbApplication.Registry.CreateApplication(
             "ConvertSensorToGeoPointFast"
         )
 
-        s2c_app.SetParameterString("in", img)
+        s2c_app.SetParameterString("in", sensor)
         s2c_app.SetParameterFloat("input.idx", x_coord)
         s2c_app.SetParameterFloat("input.idy", y_coord)
 
         if z_coord is not None:
             s2c_app.SetParameterFloat("input.idz", z_coord)
-        if dem is not None:
-            s2c_app.SetParameterString("elevation.dem", dem)
-        if geoid is not None:
-            s2c_app.SetParameterString("elevation.geoid", geoid)
-        if default_elevation is not None:
-            s2c_app.SetParameterFloat("elevation.default", default_elevation)
+        if self.dem is not None:
+            s2c_app.SetParameterString("elevation.dem", self.dem)
+        if self.geoid is not None:
+            s2c_app.SetParameterString("elevation.geoid", self.geoid)
+        if self.default_alt is not None:
+            s2c_app.SetParameterFloat("elevation.default", self.default_alt)
         # else ConvertSensorToGeoPointFast have only X, Y and OTB
 
         s2c_app.Execute()
@@ -439,26 +414,13 @@ class OTBGeometry(AbstractGeometry):
 
         return np.array([lat, lon, alt])
 
-    def image_envelope(
-        self,
-        conf,
-        product_key: str,
-        shp,
-        dem=None,
-        default_alt=None,
-        geoid=None,
-    ):
+    def image_envelope(self, sensor, geomodel, shp=None):
         """
         Export the image footprint to a shapefile
 
-        :param conf: cars input configuration dictionary
-        :param product_key: input_parameters.PRODUCT1_KEY or
-               input_parameters.PRODUCT2_KEY to identify which geometric model
-               shall be taken to perform the method
+        :param sensor: path to sensor image
+        :param geomodel: path and attributes for geometrical model
         :param shp: Path to the output shapefile
-        :param dem: Directory containing DEM tiles
-        :param default_alt: Default altitude above ellipsoid
-        :param geoid: path to geoid file
         """
         # save os env
         env_save = os.environ.copy()
@@ -471,15 +433,10 @@ class OTBGeometry(AbstractGeometry):
             del os.environ["OTB_GEOID_FILE"]
 
         # create OTB application
-        img = conf[
-            input_parameters.create_img_tag_from_product_key(product_key)
-        ]
-
-        # reset OTB DEMHandler
         loc_app = otbApplication.Registry.CreateApplication(
             "ConvertSensorToGeoPointFast"
         )
-        loc_app.SetParameterString("in", img)
+        loc_app.SetParameterString("in", sensor)
         loc_app.SetParameterFloat("input.idx", 0.0)
         loc_app.SetParameterFloat("input.idy", 0.0)
         loc_app.SetParameterFloat("input.idz", 0.0)
@@ -487,19 +444,19 @@ class OTBGeometry(AbstractGeometry):
 
         app = otbApplication.Registry.CreateApplication("ImageEnvelope")
 
-        if isinstance(img, str):
-            app.SetParameterString("in", img)
+        if isinstance(sensor, str):
+            app.SetParameterString("in", sensor)
         else:
-            app.SetParameterInputImage("in", img)
+            app.SetParameterInputImage("in", sensor)
 
-        if dem is not None:
-            app.SetParameterString("elev.dem", dem)
+        if self.dem is not None:
+            app.SetParameterString("elev.dem", self.dem)
 
-        if default_alt is not None:
-            app.SetParameterFloat("elev.default", default_alt)
+        if self.default_alt is not None:
+            app.SetParameterFloat("elev.default", self.default_alt)
 
-        if geoid is not None:
-            app.SetParameterString("elev.geoid", geoid)
+        if self.geoid is not None:
+            app.SetParameterString("elev.geoid", self.geoid)
 
         app.SetParameterString("out", shp)
         app.ExecuteAndWriteOutput()
